refactor: move unit page to spa
