+++ apps/admin-ui/src/spa/my-units/index.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import { H1 } from "common/src/common/typography";
 import { Container } from "@/styles/layout";
-import { Filters } from "@/component/Unit/Filters";
+import { Filters } from "@/spa/unit/Filters";
 import { HR } from "@/component/Table";
-import { UnitsDataLoader } from "@/component/Unit/UnitsDataLoader";
+import { UnitsDataLoader } from "@/spa/unit/UnitsDataLoader";
 
 // NOTE copy pasta from Unit/Units.tsx
 export function MyUnits() {
