+++ apps/admin-ui/src/spa/unit/Filters.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { AutoGrid, FullRow } from "@/styles/layout";
-import { SearchFilter } from "../QueryParamFilters";
-import { SearchTags } from "../SearchTags";
+import { SearchFilter } from "@/component/QueryParamFilters";
+import { SearchTags } from "@/component/SearchTags";
 
 export function Filters(): JSX.Element {
   const translateTag = (key: string, value: string) => {
