+++ apps/admin-ui/src/spa/unit/SpacesResources.tsx
@@ -1,17 +1,17 @@
 import { Button, IconPlusCircleFill } from "hds-react";
 import React from "react";
-import { useTranslation } from "react-i18next";
+import { useTranslation } from "next-i18next";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
 import { useUnitQuery } from "@gql/gql-types";
 import { Container } from "@/styles/layout";
-import Loader from "../Loader";
+import Loader from "@/component/Loader";
 import { ResourcesTable } from "./ResourcesTable";
 import { SpacesTable } from "./SpacesTable";
 import { SubPageHead } from "./SubPageHead";
-import Modal, { useModal } from "../HDSModal";
-import { NewSpaceModal } from "../Spaces/space-editor/new-space-modal/NewSpaceModal";
-import { NewResourceModal } from "../Resources/resource-editor/NewResourceModal";
+import Modal, { useModal } from "@/component/HDSModal";
+import { NewSpaceModal } from "./space-editor/new-space-modal/NewSpaceModal";
+import { NewResourceModal } from "./resource-editor/NewResourceModal";
 import { base64encode } from "common/src/helpers";
 import { errorToast } from "common/src/common/toast";
 import Error404 from "@/common/Error404";
