+++ apps/admin-ui/src/spa/unit/resource-editor/EditForm.tsx
@@ -4,7 +4,7 @@ import {
   type ResourceUpdateForm,
 } from "./modules/resourceEditor";
 import { TextInput } from "hds-react";
-import { ParentSelector } from "@/component/Spaces/space-editor/ParentSelector";
+import { ParentSelector } from "../space-editor/ParentSelector";
 import { useTranslation } from "next-i18next";
 import { getTranslatedError } from "@/common/util";
