+++ apps/admin-ui/src/spa/unit/resource-editor/ResourceEditorView.tsx
