+++ apps/admin-ui/src/spa/unit/resource-editor/modules/resourceEditor.ts
