+++ apps/admin-ui/src/spa/unit/resource-editor/queries.tsx
